requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
pip==23.3.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
Pillow==10.1.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
ultralytics==8.0.221
tensorflow==2.15.0
deprecated==1.2.14
gdown==4.7.1
gitpython==3.1.40
paho-mqtt==1.6.1
aiohttp==3.9.1
celery==5.3.6
redis==5.0.1